More tests and folder cleaning
First and foremost there is a proper package name, there are tests that check
the external-part of the sorting and there are a few additions to the argument
parser to allow 10k or 10m lines per batch (vs writing it all out). Furthermore,
all files in the temporary folder are deleted after a successful run.

There is still a lot to be desired of this program but it is usable in some
cases. I haven't done any performance monitoring though.