Robustness work + generic autoMerge + start ComplexRows
The reader now can handle faulty lines better and should just continue
operation while showing which lines were not read. Even for VaryRows it
shows which part of the encoding goes wrong.

Doing mergeAccidentalSorted used to compare files based on Doubles
but now it compares based on the type of file (type of sorting column).
For VaryRows it stays using a list of doubles. To make this happen the
FileSort lt & gt are doing in a generic way.

To handle files with very large integers and numbers there is a
ComplexVaryRow coming. This commit adds the InSort part of that.